bug in deletion

+++ src/crud/allocation.py
@@ -1,5 +1,6 @@
 from src.database import allocation_history, allocationDB_helper
 from bson.objectid import ObjectId
+from datetime import date, datetime
 
 
 #crud operations
@@ -21,7 +22,14 @@ async def add_allocation(allocation_data: dict) -> dict:
 #Delete an allocation
 async def delete_allocation(id : str):
     allocation = await allocation_history.find_one({"_id": ObjectId(id)})
+    
     if allocation:
-        await allocation_history.delete_one({"_id": ObjectId(id)})
-        return True
+        allocation_date_str = datetime.strptime(allocation["allocation_date"], "%Y-%m-%d").date()
+        if allocation_date_str >= date.today():
+            return 0  # let 0 means date it is not future date
+        else: 
+            await allocation_history.delete_one({"_id": ObjectId(id)})
+            return 1  # let 1 means successfully deleted
+    return 2  #let 2 means date doesn't exist
+
